First steps for the refactoring of the validator
